backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "A healthcare search engine"
authors = ["raahul"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
fastapi = "^0.109.0"
fastapi-doc-http-response = "^0.0.3"
fastapi-versioning = "^0.10.0"
llama-index = "^0.10.18"
phoenix = "^0.9.1"
pyvis = "^0.3.2"
requests = "^2.31.0"
sqlalchemy = "^2.0.15"
authx = "1.0.1b1"
itsdangerous = "^2.1.2"
llama-index-embeddings-text-embeddings-inference = "^0.1.2"
psycopg2-binary = "^2.9.9"
matplotlib = "^3.8.3"
python-multipart = "^0.0.9"
redis = "^5.0.3"
sentry-sdk = {extras = ["opentelemetry"], version = "^1.42.0"}
dspy = "^0.0.7"
opentelemetry-distro = {extras = ["otlp"], version = "^0.44b0"}
opentelemetry-sdk-extension-aws = "^2.0.1"
opentelemetry-propagator-aws-xray = "^1.0.1"
ruff = "^0.3.3"
nebula3-python = "^3.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
